pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "agentscope>=1.0.16",
    "fastapi>=0.116.1",
    "pycryptodome>=3.23.0",
    "pyodbc>=5.2.0",
    "python-dotenv>=1.2.2",
    "redis>=5.2.1",
    "uvicorn>=0.35.0",
]

[project.scripts]
demo-api = "app.main:main"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
